Issues #SB-5284 feat: adding test case for user login id
