+++ sunbird_service_api_test/src/test/java/org/sunbird/integration/test/user/UserTest.java
@@ -44,7 +44,6 @@ public class UserTest extends BaseCitrusTest {
   public static final String CREATE_USER_LOCAL_URI = "/v1/user/create";
   private static final String UPDATE_USER_LOCAL_URI = "/v1/user/update";
   public static final String TEMPLATE_DIR = "templates/user/create";
-  private static final String TEMPLATE_DIR_GET_USER_BY_LOGIN_ID = "templates/user/getbyloginid";
   private static volatile String USER_NAME = "userName";
   private static String externalId = String.valueOf(System.currentTimeMillis());
   private static String provider = String.valueOf(System.currentTimeMillis() + 10);
@@ -113,15 +112,6 @@ public class UserTest extends BaseCitrusTest {
         Constant.UPDATE_USER_TEMPLATE_LOCATION + "user_update_bad_request_response.json",
         "invalidRequestDataChannelTest"
       }
-    };
-  }
-
-  @DataProvider(name = "getUserByLoginIdFailure")
-  public Object[][] getUserByLoginIdFailure() {
-    return new Object[][] {
-      new Object[] {"testGetUserByLoginIdFailureWithInvalidLoginId"},
-      new Object[] {"testGetUserByLoginIdFailureWithEmptyLoginId"},
-      new Object[] {"testGetUserByLoginIdFailureWithInvalidRequest"}
     };
   }
 
@@ -341,21 +331,6 @@ public class UserTest extends BaseCitrusTest {
             });
   }
 
-  @Test(dataProvider = "getUserByLoginIdFailure")
-  @CitrusParameters({"testName"})
-  @CitrusTest
-  public void testGetUserByLoginIdFailure(String testName) {
-    performPostTest(
-        testName,
-        TEMPLATE_DIR_GET_USER_BY_LOGIN_ID,
-        getLmsApiUriPath(GET_USER_BY_ID_SERVER_URI, GET_USER_BY_ID_LOCAL_URI),
-        REQUEST_JSON,
-        HttpStatus.BAD_REQUEST,
-        RESPONSE_JSON,
-        false,
-        MediaType.APPLICATION_JSON);
-  }
-
   @Test(dependsOnMethods = {"testCreateUser"})
   @CitrusTest
   public void testGetUserByLoginIdSuccess() {
@@ -363,7 +338,7 @@ public class UserTest extends BaseCitrusTest {
     variable("channel", initGlobalValues.getSunbirdDefaultChannel());
     performPostTest(
         "testGetUserByLoginIdSuccess",
-        TEMPLATE_DIR_GET_USER_BY_LOGIN_ID,
+        GetUserByLoginIdTest.TEMPLATE_DIR,
         getLmsApiUriPath(GET_USER_BY_ID_SERVER_URI, GET_USER_BY_ID_LOCAL_URI),
         REQUEST_JSON,
         HttpStatus.OK,
