+++ sunbird_service_api_test/src/test/resources/templates/user/getbyloginid/testGetUserByLoginIdFailureWithInvalidRequest/request.json
@@ -1,5 +1,5 @@
 {
 	"request": {
-		"loginid": null
+		"loginid": "anyValue"
 	}
 }
